+++ docs/source/change_log.rst
@@ -14,8 +14,12 @@ Version 1.9.x
 
 * Support resetting plugins options to their respective defaults
 * Moved Office 2007+ metadata removal to a new plugin
-* Upgrade AdvancedHTTPServer to v2.0.11 to support async SSL handshakes
-* Support using an include directive in the server configuration file
+
+* Multiple server improvements
+
+   * Upgrade AdvancedHTTPServer to v2.0.11 to support async SSL handshakes
+   * Support using an include directive in the server configuration file
+   * Added a ``request-handle`` signal for custom HTTP request handlers
 
 Version 1.8.0
 ^^^^^^^^^^^^^